backend/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
opencv-python
numpy
pillow
requests
python-multipart